UR-2792 Enhance - Add notice for email address confirmation setting disabled and remove always active from email setting.
When email address confirmation is disabled, a notice is displayed inline similar to membership thankyou redirection notice in case of no redirection selected on form settings.
Notice will be displayed on the form builder when email settings for email address confirmation is not enabled and auto approval after email confirmation is selected.
Email address confirmation setting no longer is always active, instead a toggle is used to enable/disable the setting.